requirements.txt
pytest~=8.3.5
json
os
pdfplumber~=0.11.5
pytest-mock
pillow~=11.1.0
pyzbar~=0.1.9